requirements.txt
asgiref==3.5.0
attrs==21.4.0
coverage==6.3.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
whitenoise==6.0.0
